backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
eth-account==0.10.0
web3==6.15.1
solders==0.20.0
base58==2.1.1
pyotp==2.9.0
qrcode==7.4.2

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Anthropic
anthropic>=0.39.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Utils
python-dotenv==1.0.1
